Add fundamental support for upscaling denoisers
Adds basic infrastructure for denoisers that can also upscale for
viewport rendering, by taking advantage of the existing resolution
divider functionality.

The implementation basically tracks two resolution dividers, the normal
one and an additional one that also has a denoiser upscale factor
applied. It then uses the latter one for rendering and processing on
noisy render buffers, and the former one for any processing happening
after denoising. This has the advantage of allowing an additional
resolution divider on top of denoiser upscaling (so can also set the
pixel size to something other than 1x and that is still being
respected). The resolution divider was made into a floating point value
to allow fractional scale factors.

Pull Request: https://projects.blender.org/blender/blender/pulls/151133

+++ src/session/session.cpp
@@ -398,9 +398,9 @@ RenderWork Session::run_update_for_next_iteration()
     /* Update camera if dimensions changed for progressive render. the camera
      * knows nothing about progressive or cropped rendering, it just gets the
      * image dimensions passed in. */
-    const int resolution = render_work.resolution_divider;
-    const int width = max(1, buffer_params_.full_width / resolution);
-    const int height = max(1, buffer_params_.full_height / resolution);
+    const float resolution = render_work.resolution_divider;
+    const int width = max(1, int(buffer_params_.full_width / resolution));
+    const int height = max(1, int(buffer_params_.full_height / resolution));
 
     scene->update_camera_resolution(progress, width, height);
